feat(store-devtools): add provideStoreDevtools function (#3537)
Closes #3527
